+++ NEWS.md
@@ -1,6 +1,8 @@
 # igoR (development version)
 
 - The minimum required **R** version is now 4.1.0.
+- `igoR::states2024` replaces `states2016` with State System Membership data
+  through December 2024.
 
 # igoR 1.0.3
 
@@ -72,4 +74,3 @@ First major version of the package.
 - `igo_members()` was vectorized.
 - `igo_search_states()` was added.
 - `igo_state_membership()` was added.
-- `igoR::states2016` was added as a replacement for `cow_country_codes`.